remove js alert msgs instead use php echo

+++ add_framework.php
@@ -82,46 +82,20 @@
 	}
 	
 	//delete code starts from here
-	if(isset($_GET['delete'])){
-		?>
-		<script>
-		swal({
-		title: "Confirm",
-		text: "Delete competency framework?",
-		icon: "warning",
-		buttons: true,
-		dangerMode: true,
-		})
-		.then((willDelete) => {
-			if (willDelete) {
-				<?php
-				$id_d=$_GET['delete'];
-				$check=$DB->get_records_sql('SELECT * from mdl_competency where competencyframeworkid=?',array($id_d));
-				if($check){
-					?>
-					swal("Alert", "The competency framework cannot be deleted! Remove the mapping before framework deletion.", "info");
-					<?php
-				}
-				else{
-					$sql_delete="DELETE from mdl_competency_framework where id=$id_d";
-					$DB->execute($sql_delete);
-					?>
-					swal("Framework has been deleted!", {
-					icon: "success",
-					});
-					<?php
-				}
-				?>
-			}
-			else {
-				swal("Framework is safe!");
-			}
-		});
-		</script>
-	<?php
+	elseif(isset($_GET['delete'])){
+		$id_d=$_GET['delete'];
+		$check=$DB->get_records_sql('SELECT * from mdl_competency where competencyframeworkid=?',array($id_d));
+		if($check){
+			$delmsg = "<font color='red'><b>The competency framework cannot be deleted! Remove the mapping before framework deletion.</b></font><br />";
+		}
+		else{
+			$sql_delete="DELETE from mdl_competency_framework where id=$id_d";
+			$DB->execute($sql_delete);
+			$delmsg = "<font color='green'><b>Framework has been deleted!</b></font><br />";
+		}
+	}
 	//unset($_GET['delete']); //Not Working
 	//redirect('./add_framework.php'); //Not Working
-	}
 	// del code ends
 
 	$rec=$DB->get_records_sql('SELECT id,shortname from mdl_competency_framework');
@@ -136,7 +110,7 @@
 					<div class='col-md-2 col-sm-4 col-xs-8'>$i. $shortname1</div>
 						<div class='col-md-10 col-sm-8 col-xs-4'>
 							<a href='edit_framework.php?edit=$id' title='Edit'><img src='./img/icons/edit.png' /></a>
-							<a href='add_framework.php?delete=$id' title='Delete'><img src='./img/icons/delete.png' /></a>
+							<a href='add_framework.php?delete=$id' onClick=\"return confirm('Delete competency framework?')\" title='Delete'><img src='./img/icons/delete.png' /></a>
 						</div>
 				  </div>"; //link to edit_framework.php and delete
 			$i++;
@@ -145,6 +119,9 @@
 
 	if(isset($msg3)){
 		echo $msg3;
+	}
+	if(isset($delmsg)){
+		echo $delmsg;
 	}
 	
 	?>
